upd: update route for content ticket

+++ src/Nude.API.Models/Messages/UserMessage.cs
@@ -5,7 +5,7 @@ namespace Nude.API.Models.Messages;
 public class UserMessages : IEntity
 {
     public int Id { get; set; }
-    public int UserId { get; set; }
-    public int ChatId { get; set; }
-    public int MessageId;
+    public long UserId { get; set; }
+    public long ChatId { get; set; }
+    public long MessageId;
 }
